Add dependency on etl tool and EtlRunner for running the ETL process from the code.
There's no need to package FAT jar of etl tool, attach this jar to this project and run it as external application.
By using the etl tool dependency (currently expected to be install in local maven repo) we are able to invoke etl process directly from code via EtlRunner class.