DP-1027 removed automatic Manager initialization in DockerManager
This is done because in case there is a need to have
arquillian.xml where we want to expand system properties,
DockerManager was effectively static so it means it tried
to parse that file without properties expanded hence it failed.
These properties can be set before DataPlatformOrchestration
is started so once DockerManager is started underneath with it,
system properties will be there to be expanded.