tests - building of platform modules is simplified
Before this commit, building of modules was done only in case
there was not any war file of that module present in the
data plaform repository. If it was there, there was force flag
which would trigger Gradle build could be used, if it was there
and flag was not used, it was just skipped.

From now on, this logic is removed and Gradle process invocation
for building wars to deploy is triggered every time.

Gradle itself takes care of war building automatically. This removes
one flag on the command line to remember.

Building of modules is done recursively. In case some module to be
deployed depends on some other module in runtime, prerequisity module
is built (and deployed) as well.