Automatic removing of VirtualBox host only interfaces.
In case VirtualBox Docker machine has been running when network has
changed or when computer was suspended / in a hibernate mode, after
it is woke up, it is not possible to connect to VirtualBox anymore.

This is automatically detected from now on and Docker machine is
stopped, interface removed and whole machine is started again
automatically.