DP-1644 Removed duplicated methods from BasicSteps
We had 3 pairs of methods with the same name in BasicSteps. The only difference
between them was the amount of parameters. There was quite a mess all around in
the code since each variant of the same method was used in different places.
Since having the public methods with the same name is generally a bad practice
this duplication was removed. I had to update a number of method calls to
utilize the remaining method with the full list of parameters.