DP-1791 - user steps refactoring
Previously the huge amount of user step definitions executed the same set of
checks (get user by username, make sure user exists, etc.) This repetitive
functionality was abstracted in a single function in UsersSteps to be used
everywhere (resolveUserId). Besides: in different steps users were referred to
either via id or via username, and it was not always obvious which method is
to be used in which step without having to look into the stepdef source code.
Hereby the user reference is made universal, making underlying code to find out
what was used: a username or a uuid. Thus the same step could be written using
both username and user_id.